+++ Riot/Modules/Application/LegacyAppDelegate.m
@@ -1765,7 +1765,7 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
             
             [self configureCallManagerIfRequiredForSession:mxSession];
             
-            [self.configuration setupSettingsFor:mxSession];
+            [self.configuration setupSettingsFor:mxSession];                                    
         }
         else if (mxSession.state == MXSessionStateStoreDataReady)
         {