+++ Riot/Modules/People/Views/InviteRecentTableViewCell.h
@@ -26,6 +26,13 @@
 extern NSString *const kInviteRecentTableViewCellPreviewButtonPressed;
 
 /**
+ Action identifier used when the user pressed 'accept' button displayed on room invitation.
+ 
+ The `userInfo` dictionary contains an `MXRoom` object under the `kInviteRecentTableViewCellRoomKey` key, representing the room of the invitation.
+ */
+extern NSString *const kInviteRecentTableViewCellAcceptButtonPressed;
+
+/**
  Action identifier used when the user pressed 'decline' button displayed on room invitation.
  
  The `userInfo` dictionary contains an `MXRoom` object under the `kInviteRecentTableViewCellRoomKey` key, representing the room of the invitation.
@@ -44,6 +51,9 @@ extern NSString *const kInviteRecentTableViewCellRoomKey;
 
 @property (weak, nonatomic) IBOutlet UIButton *leftButton;
 @property (weak, nonatomic) IBOutlet UIButton *rightButton;
+
+@property (weak, nonatomic) IBOutlet UIActivityIndicatorView *leftButtonActivityIndicator;
+@property (weak, nonatomic) IBOutlet UIActivityIndicatorView *rightButtonActivityIndicator;
 
 @property (weak, nonatomic) IBOutlet UIView  *noticeBadgeView;