+++ Config/AppConfiguration.swift
@@ -31,6 +31,10 @@ class AppConfiguration: CommonConfiguration {
         // Enable CallKit for app
         MXKAppSettings.standard()?.isCallKitEnabled = true
         
+        // Get modular widget events in rooms histories
+        MXKAppSettings.standard()?.addSupportedEventTypes([kWidgetMatrixEventTypeString,
+                                                           kWidgetModularEventTypeString])
+        
         // Hide undecryptable messages that were sent while the user was not in the room
         MXKAppSettings.standard()?.hidePreJoinedUndecryptableEvents = true