+++ CHANGES.rst
@@ -9,7 +9,7 @@ Changes to be released in next version
  * Empty views: Add empty screen when there is nothing to display on home, people, favourites and rooms screen (#3836).
 
 🐛 Bugfix
- * 
+ * Restore the modular widget events in the rooms histories.
 
 ⚠️ API Changes
  * Slight API changes for SlidingModalPresenter to avoid race conditions while sharing a presenter. (#3842)