+++ CHANGES.rst
@@ -1,4 +1,4 @@
-Changes to be released in next version
+Changes in 1.1.2 (2020-12-02)
 =================================================
 
 ✨ Features
@@ -22,6 +22,9 @@ Changes to be released in next version
 
 Others
  * 
+
+Improvements:
+ * Upgrade MatrixKit version ([v0.13.2](https://github.com/matrix-org/matrix-ios-kit/releases/tag/v0.13.2)).
 
 Changes in 1.1.1 (2020-11-24)
 =================================================